PlacesPage and error in migrations

+++ backend/db/migrations/20220207175915-create-place.js
@@ -30,7 +30,7 @@ module.exports = {
       },
       price: {
         allowNull: false,
-        type: Sequelize.DECIMAL
+        type: Sequelize.INTEGER
       },
       createdAt: {
         allowNull: false,
